Added sam into the build

+++ sys/src/cmd/sam/rasp.c
@@ -105,7 +105,7 @@ raspflush(File *f)
 void
 raspdelete(File *f, uint p1, uint p2, int toterm)
 {
-	int32_t n;
+	long n;
 
 	n = p2 - p1;
 	if(n == 0)
